Title: 최소 에너지 비용, Time: 125ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250512/최소 에너지 비용/minimum-energy-cost.java
@@ -8,7 +8,7 @@ public class Main {
 
         int n = Integer.parseInt(br.readLine());
         
-        int[] moveCost = new int[n];
+        long[] moveCost = new long[n];
 
         st = new StringTokenizer(br.readLine());
         
@@ -18,7 +18,7 @@ public class Main {
 
         st = new StringTokenizer(br.readLine());
 
-        int[] energyCost = new int[n];
+        long[] energyCost = new long[n];
 
         for(int i = 0; i < n; i++)
             energyCost[i] = Integer.parseInt(st.nextToken());
